HACK: dra7-evm: fix usb id detection after suspend/resume
On dra7, the omap-gpio driver stops generating interrupts if
disable_irq_wake is called on the gpio interrupt causing us
to miss usb otg adapter changes.

We fix our USB case by not calling disable_irq_wake ever
but this must be ideally fixed in the omap-gpio driver.
That's why this is a HACK.

Cc: Nishant Menon <nm@ti.com>
Cc: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>