HACK: UAPI: Increasing VIDEO_MAX_FRAME/VB2_MAX_FRAME to 128.
This is required because on TI SoC, the IVA_HD can consume as many as
32 buffers for some H.264 streams. And each buffer can contain plane/field
data as Y-top, Y-bottom, UV-top and UV-bottom. Each field will be treated
as a separate FRAME.

This is a hack because it is not upstreamable as is.
The VIDEO_MAX_FRAME constant is also used for the older videobuf
based driver and we do not have any idea of the impact there.
VB2_MAX_FRAME was introduce earlier as a step but for now its
value must match VIDEO_MAX_FRAME, hence we are keeping them in-sync.

Please note that the buffer related internal structure in videobuf2 uses
array to keep track of buffers so this change increase their size
accordingly regardless of the fact that a specific driver/application
only needs 8 buffers for example.
There was a discussion in the upstream mailing list but nothing
concrete so far: (see provided link)
http://www.spinics.net/lists/linux-media/msg81859.html

So this change is tentative until a better solution is available.

Signed-off-by: Somnath Mukherjee <somnath@ti.com>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>