ARM: dts: am437x-gp-evm: Remove duplicate pin from unused_pins
Currently 0x154 is duplicated in unused pins and matrix_keypad_sleep
pinctrl entries in the device tree. This patch removes the duplicate
state from unused_pins so that matrix_keypad is able to be controlled
by pinctrl, otherwise the pinctrl framework fails to bind matrix_keypad
states and none of them get populated.

0x154 has a default state as spi0_d0.gpio0_3 pin which conflicts with
the pin we use for gpio wake from suspend, MUXMODE9 on 0x1ac which is
mcasp0_ahclkx.gpio0_3. Because of that we must configure
spi0_d0.gpio0_3 (0x154) away from it's default state, otherwise we
cannot wake from gpio0 because both pins are mux'd to gpio0_3 and the
state of the pin we want, 0x1ac, is undefined. Because of this, 0x154
belongs with the matrix-keypad pins.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>