ARM: OMAP2+: Remove omap_mmu_dev_attr structure
[ Upstream commit b6a89194182fe7a33d383463b8b9af6e117d8146 ]

The structure omap_mmu_dev_attr was used in the hwmod data for
supplying device-specific data through the .dev_attr field and
used in constructing the platform data for legacy device creation.
The legacy device creation of OMAP IOMMU devices has been cleaned
up, and this structure is no longer needed, so remove it.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>
[s-anna@ti.com: cherry-pick commit 'b6a89194182f' from v4.5]