ARM: dts: dra7x-evm: add ABE DPLL init setups
ABE DPLL is only present on certain configurations of DRA7 silicon.
Currently the ABE DPLL init rate is hardcoded in kernel, which causes
init delays on devices without it. Move the init declaration into
dts files for the boards that actually have it, so the kernel side
configuration can be removed.

See upstream patches from Peter:
http://www.spinics.net/lists/linux-omap/msg121479.html
http://www.spinics.net/lists/linux-omap/msg121477.html

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Cc: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>