soc: ti: qmss: make acc queue support optional in the driver
commit 045016902bf7abeeb2a86fc9284c30dce228f055 upstream.

acc channels are available only if accumulator PDSP is loaded and
running in the SoC. As this requires firmware and user may not have
firmware in the file system, make the accumulator queue support
available in qmss driver optional. To use accumulator queus user needs
to add firmware to the file system and boot up kernel.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Santosh Shilimkar <ssantosh@kernel.org>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>