ARM: dts: k2g-evm: Add a common DSP shared memory CMA pool
A common CMA memory pool reserved memory node is added, and
is attached to all the DSP nodes through the 'memory-region'
property on the K2 Galileo EVM board. This area will be used
for allocating virtio rings and buffers. The common node allows
the DSP Memory Protection and Address Extension (MPAX) module
to be configured properly for the DSP processor.

Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: correct CMA pool addresses]
Signed-off-by: Suman Anna <s-anna@ti.com>