ARM: dts: dra72-evm: Fix CMA pool creation with 64-bit address support
Commit 5c4579abff24 ("ARM: dts: DRA7: change address-cells and
size-cells") has updated the DT root node's '#address-cells' and
'#size-cells' to 2 to support boards with greater than 2 GB of
RAM through 64-bit physical address support. The CMA pools fail
to be created after the above commit due to the mismatch in the
reserved-memory node's '#address-cells' and '#size-cells'
properties, and thereby causing the OMAP remoteproc driver to
fail in turn during the probe.

The following build warnings are also seen with the above patch
indicating the potential failure,
  DTC     arch/arm/boot/dts/dra72-evm.dtb
  Warning (ranges_format): /reserved-memory has empty "ranges" property but its #address-cells (1) differs from / (2)
  Warning (ranges_format): /reserved-memory has empty "ranges" property but its #size-cells (1) differs from / (2)

So, update the #address-cells and #size-cells in reserved-memory
node to match the root node and fix the OMAP remoteproc failures.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
[s-anna@ti.com: rewrite commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>