ARM: dts: DRA7: Fix ocp address and size cells
commit 5c4579abff24a ("ARM: dts: DRA7: change address-cells and size-cells")
updates the address and size cells of the nodes directly attached to root
node to 2 inorder to support 64 bit memory address space. But it forgot to
update the ocp node because of which the following error comes during boot:

[    0.346038] omap_l3_noc 4400000001000000.ocp: invalid resource
[    0.346051] omap_l3_noc 4400000001000000.ocp: ioremap 2 failed
[    0.346077] omap_l3_noc: probe of 4400000001000000.ocp failed with error -22

Fix it by updating the address and size cells of ocp node.

Reported-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>