drm/omap: make omapdrm depend on OMAP_DSS_BASE
omapdss-base module was created some time ago, which contains the
support for the hooks omapdrm needs. omapdss-base is enabled when
OMAP_DSS_BASE Kconfig symbol is set.

However, omapdrm still depend on OMAP2_DSS, which is not correct, and
prevents from using omapdrm with the upcoming DSS6 driver. So fix the
Kconfig dependency.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>