ARM: dts: am57xx-idk-common: Add mailboxes to the PRU nodes
Add the optional 'mboxes' property to both the PRU cores in each
of PRU-ICSS1 and PRU-ICSS2 remote processor subsystems for the
AM57xx IDK boards. This is usually needed for running the Remote
Processor Messaging (RPMsg) stack between the host processor and
each of the remote processors.

Each of the PRU cores uses a single sub-mailbox node, the chosen
sub-mailboxes should match the values used in the current firmware
images. The respective PRUSS crossbar also needs to be setup properly
on the firmware side to achieve proper communication as the Mailbox
interrupts are not mapped to any of the existing events by default.
This can be changed, if needed, as per the system integration needs
after making appropriate changes on the firmware side as well.

Tested-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>