ARM: dts: AM571x-IDK Initial Support
The AM571x-IDK board is a board based on TI's AM5718 SOC
which has a single core 1.5GHz A15 processor. This board is a
development platform for the Industrial market with:
- 1GB of DDR3L
- Dual 1Gbps Ethernet
- HDMI,
- PRU-ICSS
- uSD
- 16GB eMMC
- CAN
- RS-485
- PCIe
- USB3.0
- Video Input Port
- Industrial IO port and expansion connector

The PRU/ICSS will be supported by 3rd party software for EtherCat,
Profibus and others.

The link to the data sheet and TRM can be found here:

http://www.ti.com/product/AM5718

Initial support is only for uart, i2c and the tps659037 pmic

Signed-off-by: Schuyler Patton <spatton@ti.com>
Reviewed-by: Lokesh Vutla <lokeshvutla@ti.com>