ti_config_fragments: Add defconfig build script
Add a new defconfig builder script that will take
inputs from a defconfig map file to generate
defconfigs per processor.

The script uses the map to identify the processor,
the base defconfig to use, a config file containing
the basic TI defconfig fragments and any additional
config fragments that are needed.

For example: To build am335x debug only mode the
Base defconfig is omap2plus_defconfg, the config
file is the defconfig_fragment and the am335x_only.cfg
is used to undefine OMAP and other non-am335 parts.

The output of this is a defconfig stored in the
arch/arm/configs directory so that the customer
can build the defconfig.

Signed-off-by: Dan Murphy <dmurphy@ti.com>