net: keystone: add support of subsystem register region regmap
This patch is staging to support 10gbe on Keysstone devices in
subsequent patches.

10gbe serdes phy driver needs to access the 10gbe subsystem control
register during serdes initialization.  To facilitate the shared
access of the subsystem register region between the 10gbe ethernet
driver and the serdes driver, this patch adds support of the
subsystem register region defined by a syscon node in the dts.

Although there is no shared access to the gbe subsystem register
region, using syscon for that is for the sake of consistency.

This change is backward compatible with previously released gbe
devicetree bindings.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>