HACK: firmware: ti_sci: Move mailbox ticker after the tick
For some reason, unless we move the ticker logic *after* transmit, we
are unable to send data, however, when we do that, we have to deal
with Mailbox core's queue logic which is limited in size, so we can
only send "so much" requests simultaneously(ridiculous limitation, but
that is how it is).

While at this, increase predicted timeout to an unexpected value.

NOTE: this is tagged as a hack as the rootcause is still unknown. Side
effect of this patch will be to force the system to have just one
xfer only in progress to TI SCI. This needs further investigation to
rootcause, but this should unblock further work using the ti_sci for
communication

Signed-off-by: Nishanth Menon <nm@ti.com>
Tested-by: Dave Gerlach <d-gerlach@ti.com>