ARM: dts: AM57xx IDK: Common and AM572x Add MMC and eMMC support
Adding MMC and eMMC support.

Both of the AM571 and AM572 IDK EVMs have the same card voltage
design as the Beagle X15/AM57xx GPEVM. This means that it the mmc
interface cannot be scaled to support the higher speeds. This
addition is following this commit for the X15:

47eb08a581e7a4
ARM: dts: am57xx-beagle-x15: remove UHS capabilities from MMC1 dt node
Author: Kishon Vijay Abraham I <kishon@ti.com>

Signed-off-by: Schuyler Patton <spatton@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>