OMAPDSS: add TI DSS6 driver
Add support for DSS6 IP on TI K2G SoC.

DSS6 is an evolution of the OMAP DSS (DSS2). OMAP DSS stayed quite
similar from OMAP2 to OMAP5 (including DRA7 and AM5 which have basically
the same IP as OMAP5), but in DSS6 lots of things have been
restructured.

DSS6 can still be considered as being the same family as DSS2, as the
features offered are mostly the same, and the model how things happen
and work are the same. What's changed are the registers and how the
functionalities in the IP are divided.

Thus there was not much point in trying to share the low level parts of
the driver, and instead a new driver was written from scratch.

This patch adds a new DSS driver for DSS6. The DSS2's panel and encoder
drivers and also omapdrm driver can be used with DSS6.

Note that omapfb is not supported with DSS6.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>