gpu: drm: i2c: adihdmi: componentize driver and huge reformat/cleanup
Convert the driver over the the device model component framework, making
use of the drm encoder/connector helpers. This allows adihdmi to be
dynamically selected as an external encoder for drm drivers like tilcdc
that support the DT graph binding which defines ports and remote-endpoints
to attach external encoders.

Also, this driver was modified by another developer to support audio and
tweak some settings.  Along the way it seems to have been reformatted to
4 space tabs which is hard to work with alongside the standard 8 space tabs
in the kernel coding standard. As such, this is reformatted to standard 8
space tabs so it's a bit more readable.

The component and audio support should be merged into the upstream driver
so this adihdmi driver can be removed.

Signed-off-by: Matt Porter <mporter@konsulko.com>