tty: serial: 8250: omap: convert to using cyclic transfers
By using cyclic transfers it is possible to sustain long running
DMA transfers at 3Mbit.

Note that an ugly device-tree lookup hack is implemented in order
to detect if the sDMA engine is being used. This is necessary
because the sDMA driver does not implement pause/resume correctly
and therefore requires slightly different logic for the UART
driver.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>