remoteproc: add infrastructure to support user-space loading/booting
The remoteproc infrastructure is enhanced to allow remoteproc drivers
to support userspace driven loading and booting mechanism. This is
done in preparation for a new remoteproc driver on TI Keystone2 SoCs.

A new field 'use_userspace_loader' is added to the rproc structure,
and this field is expected to be set by remoteproc drivers wishing
to bypass the in-kernel loading and boot mechanism. The remoteproc
core skips processing of any virtio devices or looking for firmware
or loading firmware segments using this state flag.

Two new API, rproc_add_vdevs_direct() and rproc_remove_vdevs_direct()
are added to allow the remoteproc drivers to trigger the processing
of the resource table. These API are expected to be invoked by the
remoteproc drivers after the loading is completed and the resource
table information is published to the remoteproc driver. The interface
and implementation details to userspace is left to the individual
remoteproc driver.

Signed-off-by: Sam Nelson <sam.nelson@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>