ARM: dts: am57xx-idk-common: Enable the system mailboxes 3 and 4
Enable the System Mailboxes 3 and 4 and the corresponding child
sub-mailbox nodes, in order to communicate with the respective
PRU remote cores in each of the PRUSS processor subsystem from
the host. It is assumed that the PRUSS crossbar is setup properly
on the PRU firmware side to receive the mailbox interrupts on one
of its input events.

Tested-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>