ARM: dts: k2g: Remove voltage-ranges property from MMC nodes
Remove voltage-ranges property from MMC nodes. The proper solution
will be to add fixed regulators that will be used by the MMC nodes.
This follows the approach suggested by Ulf Hansson in the below thread:
https://lkml.org/lkml/2015/8/3/280

The fixed regulators will be added to the board specific dts when the MMCs
are enabled in later patches.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>