TI-Integration: Move all Media/V4L2 related config to audio-display fragment
Some of the V4L2 related config in audio_display.cfg are getting
overwritten when the connectivity.cfg fragment gets merge
because some of the config are duplicated there and since it is
getting merged last then whatver we intend in the audio_display.cfg
gets thrown away.

Up to now CONFIG_MEDA_SUPPORT was set to "=m" in both fragment
so the conflict was not visible.

But with the addition of the Write Back feature as part of omapdrm,
which is configured as built-in, the needed V4L2 config must be also
be switch to buil-in.

But in the current state it gets switch back to module which basically
disabled Writeback altogether.

In order to fix these config conflict while configuring
V4L2, it is preferable if all of the V4L2 config are handled
by a single config fragment.

Now as part of merging these config into a single file
we also cleaned up (as in remove) configs which are automatically
derived form higher level selection.

These lines are removed for the following reason:

-#USB Video
-CONFIG_MEDIA_SUPPORT=m
 	Already in audio-displat.cfg
-CONFIG_MEDIA_CAMERA_SUPPORT=y
 	Already in audio-displat.cfg
-CONFIG_VIDEO_DEV=m
 	Not directly settable in menuconfig, this is derived from
	CONFIG_MEDIA_SUPPORT
-CONFIG_VIDEO_V4L2=m
 	Not directly settable in menuconfig, this is derived from
	CONFIG_MEDIA_SUPPORT
-CONFIG_VIDEOBUF2_CORE=m
 	Not directly settable in menuconfig, this is/should be
	"selected" by the requesting driver module
-CONFIG_VIDEOBUF2_MEMOPS=m
 	Not directly settable in menuconfig, this is/should be
	"selected" by the requesting driver module
-CONFIG_VIDEOBUF2_VMALLOC=m
 	Not directly settable in menuconfig, this is/should be
	"selected" by the requesting driver module
-CONFIG_MEDIA_USB_SUPPORT=y
	Move to audio-display.cfg
-CONFIG_USB_VIDEO_CLASS=m
	Move to audio-display.cfg
-CONFIG_USB_VIDEO_CLASS_INPUT_EVDEV=y
 	Looks like this are automatically selected by
	CONFIG_USB_VIDEO_CLASS
-CONFIG_USB_GSPCA=m
 	Looks like this are automatically selected by
	CONFIG_USB_VIDEO_CLASS

These were found by carefully comparing "make savedefconfig" outputs
before and after these changes.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>