ARM: dts: k2g: Fix Message Manager interrupt polarity
It seems that the message manager proxy interrupt is held high for as
long as there is data in the queue. Since we handle a single message
in the TI message manager driver, If there are multiple messages already
pending in the queue prior to the driver's handler getting invoked,
Then, only the first message is handled (triggered by the rising edge).

Other pending messages in the queue are never handled and system
deadlocks and eventually queue overflows.

Signed-off-by: Nishanth Menon <nm@ti.com>