remoteproc/keystone: add support for using reset framework
The Keystone family of SoCs has the reset registers as part of the
Power and Sleep Controller (PSC) module. The keystone remoteproc
driver currently manages these DSP resets by itself through the
regmap/syscon API.

Enhance the keystone remoteproc driver to start using the reset
framework API for managing resets. This is being done to streamline
the driver usage for supporting the DSP on Keystone 2 Galileo (K2G)
SoC. This also switches the driver to use a more standard framework
for resets.

The support is added as an incremental change, so the regmap/syscon
framework is used as a fallback to support the transition of the
Keystone 2 DSP DTS nodes from using 'ti,psc-syscon' to the 'resets'
property.

Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: add error checking, checkpatch fixes, binding updates]
Signed-off-by: Suman Anna <s-anna@ti.com>