firmware: ti_sci: Add device_resets calls to the device ops
Add set_device_resets to the device ops of ti_sci to handle new
TI_SCI_MSG_SET_DEVICE_RESETS message introduced with ABI 2.3 that
will control the resets of a device without touching any of the
other state associated with that device.

Previously resets control was handled by the TI_SCI_MSG_SET_DEVICE
message which also controls device state. This would lead to the
reset framework being responsible for controlling resets but also
requiring knowledge about the state of the device or making assumptions
and causing undefined behavior. With this new split ti_sci_pm_domains
can control the state exclusively while also setting the reset value
when turning the device on and the reset framework will be able to
control resets independently of device state.

Also add get_device_resets to return the state of the resets of a
device.

Acked-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>