mmc: host: omap_hsmmc: increase the timeout value given for sw timeout
Increase the timeout value given for sw timeout to a high value
(3 times the value requested by card). During the experiments, it
was observed that the eMMC card on DRA72x EVM advertised 960ms as
timeout but the (worst case) transfer complete came after 2.6 seconds.

The additional delay is to account for the start time of the transfer,
IO timings and card advertising incorrect timeout values. Also add a
minimum timeout value of 20ms so that we don't timeout unnecessarily.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
[nsekhar@ti.com: massage commit text]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>