TEMP: ARM: dts: k2g-evm: Reserve some DDR for MPM usecases
Add a reserved memory node to carveout a portion of the external
DDR memory, and add this address region to the dspmem node on the
K2 Galileo EVM board. This is done to enable the Keystone
Multi Proc Manager (MPM) to load images that use this memory
region. The memory is chosen to be adjacent to the DSP CMA memory
pool so that the DSP Memory Protection and Address Extension
(MPAX) module can be configured efficiently. The memory will
not be mapped into the kernel space.

Signed-off-by: Suman Anna <s-anna@ti.com>