mmc: host: omap_hsmmc: Fix OMAP_HSMMC_SETUP_PINCTRL to set the correct 'cap' variable
HS200 capability is set in 'caps2' member while the rest of the other
capabilities are set in 'caps' member of mmc_host.

OMAP_HSMMC_SETUP_PINCTRL always resets the capabilities present in
'caps' member if it's not able to find the pinctrl state of a particular
mode. However if pinctrl state of HS200 is not present, HS200 capability
should be reset in 'caps2' member.

Fixes: <c337a0486cbc> ("mmc: host: omap_hsmmc: get pinctrl states only for
the supported capabilities")

Reported-by: Yan Liu <yan-liu@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Cc: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>