media: ti-vpe: vip: Fix a minor sensor handshake issue
Currently when a vip device is opened the sensor is
asked what is current format is. But the lookup function
currently used is meant for pixelformat/fourcc lookup
and not mbus_code.
The side effect is that the lookup will always fail forcing
the vip driver to issue a set format sequence which is not
necessary.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>