ARM: boot: dts: beaglex15: Remove ID GPIO
According to the latest schematics[1] USB2_ID is
hard wired to operate on peripheral mode only and
the ID pin on the connector is left floating.

In fact, newer revisions(Rev A3+) of the platforms
will not even provision for host mode
operation.

To avoid problems in the future and skip an
unnecessary check of ID status from GPIO, let's
remove references of ID gpio from DTS.

[1] https://github.com/beagleboard/beagleboard-x15/blob/master/BeagleBoard-X15_RevA2.pdf

Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>